requirements.txt
scrapy==2.8.0
selenium==4.9.1
webdriver-manager==3.8.6
opencv-python==4.7.0.72
Pillow==9.5.0
scikit-image==0.19.3
scikit-learn==1.0.2
faiss-cpu==1.7.4
flask==2.2.5
flask-cors==4.0.0
numpy==1.21.6
pandas==1.3.5
requests==2.31.0
